Rename visitor.py to parser.py, inspect node names in docs_test.py
