+++ env-var-docs/env_var_docs/setup.py
@@ -3,6 +3,6 @@ import setuptools
 setuptools.setup(
     name="env_var_docs",
     version="1.0.0",
-    py_modules=["visitor"],
+    py_modules=["parser"],
     python_requires=">=3.10",
 )
